Backend/requirements.txt
flask
transformers
torch
langchain
faiss-cpu
PyPDF2
python-dotenv
